+++ examples/nested-loops.hs
@@ -1,15 +1,14 @@
 import Control.Concurrent (myThreadId)
-import Control.Monad.IO.Class (liftIO)
 import System.IO (stdout, hSetBuffering, BufferMode(LineBuffering))
 import System.Random (randomIO)
 import Streamly
 import Streamly.Prelude (nil)
 
 main = runStream $ do
-    liftIO $ hSetBuffering stdout LineBuffering
+    fromIO $ hSetBuffering stdout LineBuffering
     x <- loop "A " 2
     y <- loop "B " 2
-    liftIO $ myThreadId >>= putStr . show
+    fromIO $ myThreadId >>= putStr . show
              >> putStr " "
              >> print (x, y)
 
@@ -17,7 +16,7 @@ main = runStream $ do
 
     loop :: String -> Int -> StreamT IO String
     loop name n = do
-        rnd <- liftIO (randomIO :: IO Int)
+        rnd <- fromIO (randomIO :: IO Int)
         let result = (name ++ show rnd)
             repeat = if n > 1 then loop name (n - 1) else nil
          in (return result) `parallel` repeat