+++ examples/parallel-loops.hs
@@ -1,20 +1,19 @@
 import Control.Concurrent (myThreadId, threadDelay)
-import Control.Monad.IO.Class (liftIO)
 import System.IO (stdout, hSetBuffering, BufferMode(LineBuffering))
 import System.Random (randomIO)
 import Streamly
 
 main = runStream $ do
-    liftIO $ hSetBuffering stdout LineBuffering
+    fromIO $ hSetBuffering stdout LineBuffering
     x <- loop "A" `parallel` loop "B"
-    liftIO $ myThreadId >>= putStr . show
+    fromIO $ myThreadId >>= putStr . show
              >> putStr " "
              >> print x
 
     where
 
-    loop :: String -> StreamT IO (String, Int)
+    loop :: String -> Stream (String, Int)
     loop name = do
-        liftIO $ threadDelay 1000000
-        rnd <- liftIO (randomIO :: IO Int)
+        fromIO $ threadDelay 1000000
+        rnd <- fromIO (randomIO :: IO Int)
         return (name, rnd) `parallel` loop name