+++ src/Streamly/Examples/AcidRainGame.hs
@@ -14,7 +14,7 @@ import Data.Semigroup (cycle1)
 data Event = Harm Int | Heal Int | Quit deriving (Show)
 
 userAction :: MonadIO m => StreamT m Event
-userAction = cycle1 $ liftIO askUser
+userAction = cycle1 $ fromIO askUser
     where
     askUser = do
         command <- getLine
@@ -24,7 +24,7 @@ userAction = cycle1 $ liftIO askUser
             _        -> putStrLn "What?" >> askUser
 
 acidRain :: MonadIO m => StreamT m Event
-acidRain = cycle1 $ liftIO (threadDelay 1000000) >> return (Harm 1)
+acidRain = cycle1 $ fromIO (threadDelay 1000000) >> return (Harm 1)
 
 game :: (MonadParallel m, MonadState Int m) => StreamT m ()
 game = do
@@ -36,7 +36,7 @@ game = do
 
     h <- get
     when (h <= 0) $ fail "You die!"
-    liftIO $ putStrLn $ "Health = " ++ show h
+    fromIO $ putStrLn $ "Health = " ++ show h
 
 acidRainGame :: IO ()
 acidRainGame = do