+++ src/Streamly/Examples/MergeSortedStreams.hs
@@ -8,26 +8,26 @@ import Data.List (sort)
 import Streamly
 import qualified Streamly.Prelude as A
 
-getSorted :: MonadIO m => StreamT m Word16
+getSorted :: Stream Word16
 getSorted = do
-    g <- liftIO getStdGen
+    g <- fromIO getStdGen
     let ls = take 100000 (randoms g) :: [Word16]
-    foldMapWith (<>) return (sort ls)
+    foldMap return (sort ls)
 
-mergeAsync :: (Ord a, MonadParallel m)
-    => StreamT m a -> StreamT m a -> StreamT m a
+-- | merge two streams generating the elements from each in parallel
+mergeAsync :: Ord a => Stream a -> Stream a -> Stream a
 mergeAsync a b = do
-    x <- lift $ async a
-    y <- lift $ async b
+    x <- fromIO $ async a
+    y <- fromIO $ async b
     merge x y
 
-merge :: (Ord a, MonadParallel m) => StreamT m a -> StreamT m a -> StreamT m a
+merge :: Ord a => Stream a -> Stream a -> Stream a
 merge a b = do
-    a1 <- lift $ A.uncons a
+    a1 <- fromIO $ A.uncons a
     case a1 of
         Nothing -> b
         Just (x, ma) -> do
-            b1 <- lift $ A.uncons b
+            b1 <- fromIO $ A.uncons b
             case b1 of
                 Nothing -> return x <> ma
                 Just (y, mb) ->